requirements.txt
ccxt==4.2.14
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
ta==0.11.0
highcharts-core==1.3.1
jinja2==3.1.2
aiohttp==3.9.1
asyncio==3.4.3
flask==3.0.0
pytest==8.0.0
pytest-asyncio==0.23.5
supervisor==4.2.5
httpx
pandas_ta
